Prepare adding Tika test documents
